Add translations & extract to hooks folder
